refactor(e2e-support): densify broad-sweep dismiss mirror for #4195
Extract broad-sweep dismiss counter, priority, hit-testable, and pop-route
groups into test/support harness modules so the suite drops to 63 physical
lines ahead of the 400/500 CI ratchet.
